divided the vode into modules
