+++ getColorFromPairNumber.js
@@ -1,0 +1,12 @@
+const MajorColors=require("./majorColors.js");
+const MinorColors=require("./minorColors.js");
+function GetColorFromPairNumber(pairNumber) {
+    let colorPair = {};
+    const zeroBasedPairNumber = pairNumber - 1;
+    const majorColorIndex =
+        Math.floor(zeroBasedPairNumber / MinorColors.length);
+    colorPair.major = MajorColors[majorColorIndex];
+    colorPair.minor = MinorColors[zeroBasedPairNumber % MinorColors.length];
+    return colorPair
+}
+module.exports=GetColorFromPairNumber;
