+++ getPairNumberFromColor.js
@@ -1,0 +1,19 @@
+const MajorColors=require("./majorColors.js");
+const MinorColors=require("./minorColors.js");
+
+function GetPairNumberFromColors(majorColor, minorColor) {
+    let majorIndex = 0;
+    let minorIndex = 0;
+    for(majorIndex = 0; majorIndex < MajorColors.length; majorIndex++) {
+        if(MajorColors[majorIndex] == majorColor) {
+            break;
+        }
+    }
+    for(minorIndex = 0; minorIndex < MinorColors.length; minorIndex++) {
+        if(MinorColors[minorIndex] == minorColor) {
+            break;
+        }
+    }
+    return majorIndex * MinorColors.length + minorIndex + 1;
+}
+module.exports=GetPairNumberFromColors;
