+++ majorColors.js
@@ -1,0 +1,9 @@
+const MajorColors = [
+    "white",
+    "red",
+    "black",
+    "yellow",
+    "violet"
+];
+
+module.exports=MajorColors;
