+++ minorColors.js
@@ -1,0 +1,8 @@
+const MinorColors = [
+    "blue",
+    "orange",
+    "green",
+    "brown",
+    "slate"
+];
+module.exports=MinorColors;
